--- tests/post_training/model_scope.json
@@ -1,232 +1,0 @@
-{
-    "vgg11": {
-        "model_name": "vgg11",
-        "quantization_params": {},
-        "metrics": {
-            "FP32 top 1": 0.68344
-        }
-    },
-    "resnet18": {
-        "model_name": "resnet18",
-        "quantization_params": {},
-        "metrics": {
-            "FP32 top 1": 0.70104
-        }
-    },
-    "mobilenetv2_050": {
-        "model_name": "mobilenetv2_050",
-        "quantization_params": {
-            "preset": "MIXED"
-        },
-        "metrics": {
-            "FP32 top 1": 0.64568
-        }
-    },
-    "mobilenetv2_050_BC": {
-        "model_name": "mobilenetv2_050",
-        "quantization_params": {
-            "preset": "MIXED",
-            "fast_bias_correction": false
-        },
-        "metrics": {
-            "FP32 top 1": 0.64568
-        }
-    },
-    "densenet121": {
-        "skipped": "temporary excluded due to memory leaks",
-        "model_name": "densenet121",
-        "quantization_params": {
-            "fast_bias_correction": false
-        }
-    },
-    "densenet121_BC": {
-        "skipped": "temporary excluded due to memory leaks",
-        "model_name": "densenet121",
-        "quantization_params": {
-            "fast_bias_correction": false
-        }
-    },
-    "tf_inception_v3": {
-        "model_name": "tf_inception_v3",
-        "quantization_params": {
-            "preset": "MIXED"
-        },
-        "metrics": {
-            "FP32 top 1": 0.77728
-        }
-    },
-    "efficientnet_b0": {
-        "model_name": "efficientnet_b0",
-        "quantization_params": {
-            "preset": "MIXED"
-        },
-        "metrics": {
-            "FP32 top 1": 0.77292
-        }
-    },
-    "efficientnet_b0_BC": {
-        "model_name": "efficientnet_b0",
-        "quantization_params": {
-            "preset": "MIXED",
-            "fast_bias_correction": false
-        },
-        "metrics": {
-            "FP32 top 1": 0.77292
-        }
-    },
-    "darknet53": {
-        "model_name": "darknet53",
-        "quantization_params": {
-            "preset": "MIXED"
-        },
-        "metrics": {
-            "FP32 top 1": 0.79858
-        }
-    },
-    "resnest14d": {
-        "model_name": "resnest14d",
-        "quantization_params": {
-            "preset": "MIXED"
-        },
-        "metrics": {
-            "FP32 top 1": 0.74862
-        }
-    },
-    "inception_resnet_v2": {
-        "model_name": "inception_resnet_v2",
-        "quantization_params": {},
-        "metrics": {
-            "FP32 top 1": 0.80102
-        }
-    },
-    "wide_resnet50_2": {
-        "model_name": "wide_resnet50_2",
-        "quantization_params": {
-            "preset": "MIXED"
-        },
-        "metrics": {
-            "FP32 top 1": 0.81414
-        }
-    },
-    "regnetx_002": {
-        "model_name": "regnetx_002",
-        "quantization_params": {
-            "preset": "MIXED"
-        },
-        "metrics": {
-            "FP32 top 1": 0.67826
-        }
-    },
-    "mobilenetv3_small_050": {
-        "model_name": "mobilenetv3_small_050",
-        "quantization_params": {
-            "preset": "MIXED"
-        },
-        "metrics": {
-            "FP32 top 1": 0.56338
-        }
-    },
-    "levit_128": {
-        "skipped_by_backends": {
-            "TORCH": "Wrong quantization config, skipped as the model could be quantized by TORCH PTQ and TORCH backend will be deprecated eventially"
-        },
-        "model_name": "levit_128",
-        "quantization_params": {
-            "preset": "MIXED",
-            "model_type": "TRANSFORMER"
-        },
-        "metrics": {
-            "FP32 top 1": 0.7405
-        }
-    },
-    "deit3_small_patch16_224": {
-        "model_name": "deit3_small_patch16_224",
-        "quantization_params": {
-            "preset": "MIXED",
-            "model_type": "TRANSFORMER"
-        },
-        "metrics": {
-            "FP32 top 1": 0.76974
-        }
-    },
-    "swin_base_patch4_window7_224": {
-        "model_name": "swin_base_patch4_window7_224",
-        "quantization_params": {
-            "preset": "MIXED",
-            "model_type": "TRANSFORMER"
-        },
-        "metrics": {
-            "FP32 top 1": 0.81462
-        }
-    },
-    "convit_tiny": {
-        "skipped": "Suppressed due to bug - 104173",
-        "model_name": "convit_tiny",
-        "quantization_params": {
-            "preset": "MIXED",
-            "model_type": "TRANSFORMER"
-        }
-    },
-    "visformer_small": {
-        "skipped_by_backends": {
-            "TORCH": "Wrong quantization config, skipped as the model could be quantized by TORCH PTQ and TORCH backend will be deprecated eventially"
-        },
-        "model_name": "visformer_small",
-        "quantization_params": {
-            "preset": "MIXED",
-            "model_type": "TRANSFORMER"
-        },
-        "metrics": {
-            "FP32 top 1": 0.77902
-        }
-    },
-    "crossvit_9_240": {
-        "skipped_by_backends": {
-            "TORCH": "Wrong quantization config, skipped as the model could be quantized by TORCH PTQ and TORCH backend will be deprecated eventially"
-        },
-        "model_name": "crossvit_9_240",
-        "quantization_params": {
-            "preset": "MIXED",
-            "model_type": "TRANSFORMER"
-        },
-        "metrics": {
-            "FP32 top 1": 0.69966
-        }
-    },
-    "hrnet_w18": {
-        "model_name": "hrnet_w18",
-        "quantization_params": {
-            "preset": "MIXED"
-        },
-        "metrics": {
-            "FP32 top 1": 0.77504
-        }
-    },
-    "efficientnet_lite0": {
-        "model_name": "efficientnet_lite0",
-        "quantization_params": {
-            "preset": "MIXED"
-        },
-        "metrics": {
-            "FP32 top 1": 0.7496
-        }
-    },
-    "dpn68": {
-        "model_name": "dpn68",
-        "quantization_params": {
-            "preset": "MIXED"
-        },
-        "metrics": {
-            "FP32 top 1": 0.75964
-        }
-    },
-    "dla34": {
-        "model_name": "dla34",
-        "quantization_params": {
-            "preset": "MIXED"
-        },
-        "metrics": {
-            "FP32 top 1": 0.74416
-        }
-    }
-}
