--- tests/post_training_quantization/requirements.txt
@@ -1,19 +1,0 @@
---extra-index-url https://download.pytorch.org/whl/cpu
-torch==2.0.1
---extra-index-url https://download.pytorch.org/whl/cpu
-torchvision==0.15.2
-transformers==4.29.2
-onnx==1.13.1
-onnxruntime==1.14.1
-pytest
-openvino-dev==2023.0.0
-optimum[onnxruntime,openvino]==1.8.8
-optimum-intel @ git+https://github.com/huggingface/optimum-intel@47428484598986443686b7c8ba825b4a3d7b4f73
-soundfile==0.12.1
-librosa==0.10.0
-memory-profiler==0.61.0
-pandas==1.3.5
-tensorboard==2.13.0
-tensorflow-io==0.32.0
-timm==0.9.2
-scikit-learn==1.2.2
